avantlush_backend/api/requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
psycopg2-binary==2.9.9
drf-yasg==1.21.7
django-debug-toolbar==4.2.0
sentry-sdk==1.39.1
